src/agents/trading_system/requirements.txt
uagents>=0.6.0
ccxt>=4.0.0
vaderSentiment>=3.3.2
pandas>=1.3.0
numpy>=1.20.0
ta>=0.10.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
websockets>=10.0
pydantic>=1.9.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0
plotly>=5.13.0
requests>=2.28.0
beautifulsoup4>=4.11.0
